Move group trigger tests to their own file
